+++ .idea/dictionaries/project.xml
@@ -4,6 +4,7 @@
       <w>elantra</w>
       <w>hangalito</w>
       <w>hangalo</w>
+      <w>litestore</w>
     </words>
   </dictionary>
 </component>
