--- src/main/java/dev/hangalito/annotations/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Library-specific annotations.
- */
-package dev.hangalito.annotations;
