--- src/main/java/dev/hangalito/exceptions/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Library-specific exceptions.
- */
-package dev.hangalito.exceptions;
