+++ src/main/java/dev/hangalito/litestore/Datasource.java
@@ -1,8 +1,8 @@
-package dev.hangalito.storage;
+package dev.hangalito.litestore;
 
-import dev.hangalito.annotations.Key;
-import dev.hangalito.exceptions.NoSuchIndexException;
-import dev.hangalito.exceptions.UnsupportedStorageException;
+import dev.hangalito.litestore.annotations.Key;
+import dev.hangalito.litestore.exceptions.NoSuchIndexException;
+import dev.hangalito.litestore.exceptions.UnsupportedStorageException;
 
 import java.io.*;
 import java.lang.reflect.Field;
@@ -377,4 +377,3 @@ public class Datasource<T extends Serializable, ID extends Serializable & Compar
     }
 
 }
-
