+++ src/main/java/dev/hangalito/litestore/annotations/Key.java
@@ -1,4 +1,4 @@
-package dev.hangalito.annotations;
+package dev.hangalito.litestore.annotations;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -16,4 +16,3 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Retention(RUNTIME)
 public @interface Key {
 }
-
