+++ src/main/java/dev/hangalito/litestore/annotations/Storable.java
@@ -1,4 +1,4 @@
-package dev.hangalito.annotations;
+package dev.hangalito.litestore.annotations;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -37,4 +37,3 @@ public @interface Storable {
      */
     String value() default "";
 }
-
