+++ src/main/java/dev/hangalito/litestore/annotations/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Library-specific annotations.
+ */
+package dev.hangalito.litestore.annotations;
