+++ src/main/java/dev/hangalito/litestore/exceptions/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Library-specific exceptions.
+ */
+package dev.hangalito.litestore.exceptions;
