+++ src/main/java/dev/hangalito/litestore/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Core classes and interfaces of the library.
  */
-package dev.hangalito.storage;
+package dev.hangalito.litestore;
